cleaning up plugins
