+++ TrackingTools/TrackFitters/plugins/module.cc
@@ -1,7 +1,5 @@
 #include "TrackingTools/TrackFitters/plugins/KFTrajectoryFitterESProducer.h" 
 #include "TrackingTools/TrackFitters/plugins/KFTrajectorySmootherESProducer.h" 
-#include "TrackingTools/TrackFitters/plugins/KFFittingSmootherESProducer.h" 
-#include "TrackingTools/TrackFitters/plugins/FlexibleKFFittingSmootherESProducer.h" 
 
 
 #include "FWCore/Framework/interface/EventSetup.h"
@@ -12,5 +10,3 @@
 
 DEFINE_FWK_EVENTSETUP_MODULE(KFTrajectoryFitterESProducer);
 DEFINE_FWK_EVENTSETUP_MODULE(KFTrajectorySmootherESProducer);
-DEFINE_FWK_EVENTSETUP_MODULE(KFFittingSmootherESProducer);
-DEFINE_FWK_EVENTSETUP_MODULE(FlexibleKFFittingSmootherESProducer);
